Make the topology control bar its own component
remove unused children prop from Topology Graph component.
